Moved some things to utils.h
