+++ src/sparsebase/format/format.h
@@ -27,136 +27,6 @@ namespace sparsebase {
 
 namespace format {
 
-//TODO: move to utils
-// Thanks to artificial mind blog: https://artificial-mind.net/blog/2020/10/03/always-false
-template <typename ... T>
-constexpr bool always_false = false;
-
-#include <limits>
-#include <stdint.h>
-
-using std::numeric_limits;
-
-// Cross float-integral type conversion is not currently available
-template <typename T, typename U>
-bool CanTypeFitValue(const U value) {
-  if constexpr (std::is_integral_v<T> != std::is_integral_v<U>) return false;
-  if constexpr (std::is_integral_v<T> && std::is_integral_v<U> ) {
-    const intmax_t botT = []() {
-      if constexpr (std::is_floating_point_v<T>)
-        return intmax_t(-(numeric_limits<T>::max()));
-      else
-        return intmax_t(numeric_limits<T>::min());
-    }();
-    const intmax_t botU = []() {
-      if constexpr (std::is_floating_point_v<U>)
-        return intmax_t(-(numeric_limits<U>::max()));
-      else
-        return intmax_t(numeric_limits<U>::min());
-    }();
-    const uintmax_t topT = uintmax_t(numeric_limits<T>::max());
-    const uintmax_t topU = uintmax_t(numeric_limits<U>::max());
-    return !((botT > botU && value < (U)(botT)) ||
-             (topT < topU && value > (U)(topT)));
-  } else if constexpr(!std::is_integral_v<T> && !std::is_integral_v<U> ) {
-    const double botT = []() {
-      if constexpr (std::is_floating_point_v<T>)
-        return T(-(numeric_limits<T>::max()));
-      else
-        return T(numeric_limits<T>::min());
-    }();
-    const double botU = []() {
-      if constexpr (std::is_floating_point_v<U>)
-        return U(-(numeric_limits<U>::max()));
-      else
-        return U(numeric_limits<U>::min());
-    }();
-    const double topT = numeric_limits<T>::max();
-    const double topU = numeric_limits<U>::max();
-    return !((botT > botU && value < (U)(botT)) ||
-             (topT < topU && value > (U)(topT)));
-  }
-  //} else if constexpr (std::is_integral_v<T> && !std::is_integral_v<U> ){
-  //  const double topT = double(numeric_limits<T>::max());
-  //  const uintmax_t topU = uintmax_t(numeric_limits<U>::max());
-  //  const double botT = []() {
-  //    if constexpr (std::is_floating_point_v<T>)
-  //      return double(-(numeric_limits<T>::max()));
-  //    else
-  //      return double(numeric_limits<T>::min());
-  //  }();
-  //  const intmax_t botU = []() {
-  //    if constexpr (std::is_floating_point_v<U>)
-  //      return intmax_t(-(numeric_limits<U>::max()));
-  //    else
-  //      return intmax_t(numeric_limits<U>::min());
-  //  }();
-  //  return !(double(topU) > topT && double(value) > topT) || !(double(botU) < botT && double(value) > topT);
-  //} else {
-  //  const uintmax_t topT = uintmax_t(numeric_limits<T>::max());
-  //  const double topU = double(numeric_limits<U>::max());
-  //  const intmax_t botT = []() {
-  //    if constexpr (std::is_floating_point_v<T>)
-  //      return intmax_t(-(numeric_limits<T>::max()));
-  //    else
-  //      return intmax_t(numeric_limits<T>::min());
-  //  }();
-  //  const double botU = []() {
-  //    if constexpr (std::is_floating_point_v<U>)
-  //      return double(-(numeric_limits<U>::max()));
-  //    else
-  //      return double(numeric_limits<U>::min());
-  //  }();
-  //  return !(topU > double(topT) && value > double(topT)) || !(botU < double(botT) && value > double(topT));
-  //}
-}
-
-//TODO: move to utils
-template <typename FromType, typename ToType>
-inline bool isTypeConversionSafe(FromType from_val, ToType to_val) {
-  return from_val == to_val && CanTypeFitValue<ToType>(from_val);
-}
-
-template <typename ToType, typename FromType, typename SizeType>
-ToType *ConvertArrayType(FromType *from_ptr, SizeType size) {
-  if (from_ptr == nullptr) return nullptr;
-  auto to_ptr = new ToType[size];
-  for (SizeType i = 0; i < size; i++){
-    to_ptr[i] = from_ptr[i];
-    if (!isTypeConversionSafe(from_ptr[i], to_ptr[i])) {
-      throw utils::TypeException(
-          "Could not convert array from type " +
-          std::string(std::type_index(typeid(FromType)).name()) + " to type " +
-          std::string(std::type_index(typeid(ToType)).name()) + ". Overflow detected");
-    }
-  }
-  return to_ptr;
-}
-
-template <typename T>
-class OnceSettable {
-public:
-  OnceSettable(): is_set_(false){}
-  operator T() const {
-    return data_;
-  }
-  OnceSettable(const OnceSettable&) = delete;
-  OnceSettable(OnceSettable&&) = delete;
-  OnceSettable& operator=(T&& data){
-    if (!is_set_) {
-      data_ = std::move(data);
-      is_set_ = true;
-      return *this;
-    }
-    throw utils::AttemptToReset<T>();
-  }
-  const T& get() const{
-    return data_;
-  }
-private:
-  T data_;
-  bool is_set_;
-};
 
 //! Enum depicting the ownership status of a Format instance
 enum Ownership {
@@ -311,7 +181,7 @@ protected:
   DimensionType order_;
   std::vector<DimensionType> dimension_;
   DimensionType nnz_;
-  OnceSettable<std::unique_ptr<sparsebase::context::Context>> context_;
+  utils::OnceSettable<std::unique_ptr<sparsebase::context::Context>> context_;
 };
 
 template <typename ValueType>
@@ -335,7 +205,7 @@ class FormatOrderOne : public FormatImplementation<FormatOrderOne<ValueType>, Fo
     template <template <typename> typename ToType, typename ToValueType>
     struct TypeConverter {
       ToType<ToValueType> * operator()(FormatOrderOne<ValueType>*, bool){
-        static_assert(always_false<ToValueType>, "Cannot do type conversion for the requested type. Throw a rock through one of our devs' windows");
+        static_assert(utils::always_false<ToValueType>, "Cannot do type conversion for the requested type. Throw a rock through one of our devs' windows");
       }
     };
 };
@@ -364,7 +234,7 @@ class FormatOrderTwo
       typename ToIDType, typename ToNNZType, typename ToValueType>
   struct TypeConverter{
     ToType<ToIDType, ToNNZType, ToValueType>* operator()(FormatOrderTwo<IDType, NNZType, ValueType>*, bool){
-      static_assert(always_false<ToIDType>, "Cannot do type conversion for the requested type. Throw a rock through one of our devs' windows");
+      static_assert(utils::always_false<ToIDType>, "Cannot do type conversion for the requested type. Throw a rock through one of our devs' windows");
     }
   };
 
@@ -557,7 +427,7 @@ struct format::FormatOrderTwo<IDType, NNZType, ValueType>::TypeConverter
 
     if (!is_move_conversion || !std::is_same_v<NNZType, ToNNZType>) {
       new_row_ptr =
-          ConvertArrayType<ToNNZType>(csr->get_row_ptr(), dims[0] + 1);
+          utils::ConvertArrayType<ToNNZType>(csr->get_row_ptr(), dims[0] + 1);
     } else {
       if constexpr (std::is_same_v<NNZType, ToNNZType>) {
         new_row_ptr = csr->release_row_ptr();
@@ -565,14 +435,14 @@ struct format::FormatOrderTwo<IDType, NNZType, ValueType>::TypeConverter
     }
 
     if (!is_move_conversion || !std::is_same_v<IDType, ToIDType>) {
-      new_col = ConvertArrayType<ToIDType>(csr->get_col(), num_nnz);
+      new_col = utils::ConvertArrayType<ToIDType>(csr->get_col(), num_nnz);
     } else {
       if constexpr (std::is_same_v<IDType, ToIDType>) {
         new_col = csr->release_col();
       }
     }
     if (!is_move_conversion || !std::is_same_v<ValueType, ToValueType>) {
-      new_vals = ConvertArrayType<ToValueType>(csr->get_vals(), num_nnz);
+      new_vals = utils::ConvertArrayType<ToValueType>(csr->get_vals(), num_nnz);
     } else {
       if constexpr (std::is_same_v<ValueType, ToValueType>) {
         new_vals = csr->release_vals();
@@ -595,20 +465,20 @@ struct format::FormatOrderTwo<IDType, NNZType, ValueType>::TypeConverter
 
     if (!is_move_conversion || !std::is_same_v<NNZType, ToNNZType>) {
       new_col_ptr =
-          ConvertArrayType<ToNNZType>(csc->get_col_ptr(), dims[0] + 1);
+          utils::ConvertArrayType<ToNNZType>(csc->get_col_ptr(), dims[0] + 1);
     } else {
       if constexpr (std::is_same_v<NNZType, ToNNZType>)
         new_col_ptr = csc->release_col_ptr();
     }
 
     if (!is_move_conversion || !std::is_same_v<IDType, ToIDType>) {
-      new_row = ConvertArrayType<ToIDType>(csc->get_row(), num_nnz);
+      new_row = utils::ConvertArrayType<ToIDType>(csc->get_row(), num_nnz);
     } else {
       if constexpr (std::is_same_v<IDType, ToIDType>)
         new_row = csc->release_row();
     }
     if (!is_move_conversion || !std::is_same_v<ValueType, ToValueType>) {
-      new_vals = ConvertArrayType<ToValueType>(csc->get_vals(), num_nnz);
+      new_vals = utils::ConvertArrayType<ToValueType>(csc->get_vals(), num_nnz);
     } else {
       if constexpr (std::is_same_v<ValueType, ToValueType>)
         new_vals = csc->release_vals();
@@ -629,7 +499,7 @@ struct format::FormatOrderTwo<IDType, NNZType, ValueType>::TypeConverter
     ToValueType * new_vals;
 
     if (!is_move_conversion || !std::is_same_v<IDType, ToIDType>) {
-      new_col = ConvertArrayType<ToIDType>(coo->get_col(), num_nnz);
+      new_col = utils::ConvertArrayType<ToIDType>(coo->get_col(), num_nnz);
     } else {
       if constexpr (std::is_same_v<IDType, ToIDType>){
         new_col = coo->release_col();
@@ -637,7 +507,7 @@ struct format::FormatOrderTwo<IDType, NNZType, ValueType>::TypeConverter
     }
 
     if (!is_move_conversion || !std::is_same_v<IDType, ToIDType>) {
-      new_row = ConvertArrayType<ToIDType>(coo->get_row(), num_nnz);
+      new_row = utils::ConvertArrayType<ToIDType>(coo->get_row(), num_nnz);
     }else {
       if constexpr (std::is_same_v<IDType, ToIDType>){
         new_row = coo->release_row();
@@ -645,7 +515,7 @@ struct format::FormatOrderTwo<IDType, NNZType, ValueType>::TypeConverter
     }
 
     if (!is_move_conversion || !std::is_same_v<ValueType, ToValueType>) {
-      new_vals = ConvertArrayType<ToValueType>(coo->get_vals(), num_nnz);
+      new_vals = utils::ConvertArrayType<ToValueType>(coo->get_vals(), num_nnz);
     } else {
       if constexpr (std::is_same_v<ValueType, ToValueType>){
         new_vals = coo->release_vals();
@@ -663,7 +533,7 @@ struct FormatOrderOne<ValueType>::TypeConverter<format::Array, ToValueType> {
     auto num_nnz = arr->get_num_nnz();
     ToValueType * new_vals;
     if (!is_move_conversion || !std::is_same_v<ValueType, ToValueType>) {
-      new_vals = ConvertArrayType<ToValueType>(arr->get_vals(), num_nnz);
+      new_vals = utils::ConvertArrayType<ToValueType>(arr->get_vals(), num_nnz);
     } else {
       if constexpr (std::is_same_v<ValueType, ToValueType>){
         new_vals = arr->release_vals();
