+++ tests/suites/sparsebase/utils/converter/converter_tests.cc
@@ -313,13 +313,13 @@ TEST(Converter, ClearingASingleDirection){
 }
 
 #define MATCH_CHECKING_ONEVAL(T1, T2, val) \
-  EXPECT_TRUE(format::isTypeConversionSafe((T1)(val), (T2)(val))); \
-  EXPECT_TRUE(format::isTypeConversionSafe((T2)(val), (T1)(val))); \
+  EXPECT_TRUE(utils::isTypeConversionSafe((T1)(val), (T2)(val))); \
+  EXPECT_TRUE(utils::isTypeConversionSafe((T2)(val), (T1)(val))); \
   if ((val)!=0) {                           \
-    EXPECT_FALSE(format::isTypeConversionSafe((T1)(-(val)), (T2)(val)));  \
-  EXPECT_FALSE(format::isTypeConversionSafe((T1)((val)), (T2)(-(val)))); \
-  EXPECT_FALSE(format::isTypeConversionSafe((T2)(-(val)), (T1)(val)));  \
-  EXPECT_FALSE(format::isTypeConversionSafe((T2)((val)), (T1)(-(val))));  \
+    EXPECT_FALSE(utils::isTypeConversionSafe((T1)(-(val)), (T2)(val)));  \
+  EXPECT_FALSE(utils::isTypeConversionSafe((T1)((val)), (T2)(-(val)))); \
+  EXPECT_FALSE(utils::isTypeConversionSafe((T2)(-(val)), (T1)(val)));  \
+  EXPECT_FALSE(utils::isTypeConversionSafe((T2)((val)), (T1)(-(val))));  \
 }
 #define MATCH_CHECKING_SIGNED_MISMATCH(T1, T2) \
   EXPECT_FALSE(format::isTypeConversionSafe((T2)(-1), (T1)(-1))); \
@@ -345,48 +345,48 @@ TEST(Converter, ClearingASingleDirection){
   if constexpr (std::is_integral_v<T1>) { \
     if constexpr (std::is_signed_v<T1>) {  \
       if constexpr (std::is_integral_v<T2>) {                                 \
-        if constexpr (std::is_signed_v<T2>) EXPECT_TRUE(format::isTypeConversionSafe((T1)(-1), (T2)(-1))); \
-        else EXPECT_FALSE(format::isTypeConversionSafe((T1)(-1), (T2)(-1)));                             \
+        if constexpr (std::is_signed_v<T2>) EXPECT_TRUE(utils::isTypeConversionSafe((T1)(-1), (T2)(-1))); \
+        else EXPECT_FALSE(utils::isTypeConversionSafe((T1)(-1), (T2)(-1)));                             \
       } else {                          \
-        EXPECT_FALSE(format::isTypeConversionSafe((T1)(2), (T2)(2))); \
-        EXPECT_FALSE(format::isTypeConversionSafe((T1)(-2), (T2)(-2))); \
+        EXPECT_FALSE(utils::isTypeConversionSafe((T1)(2), (T2)(2))); \
+        EXPECT_FALSE(utils::isTypeConversionSafe((T1)(-2), (T2)(-2))); \
       }                               \
     } else {                          \
         \
     }                                    \
   } else {                            \
     if constexpr (std::is_integral_v<T2>){\
-        EXPECT_FALSE(format::isTypeConversionSafe((T1)(1.5), (T2)(1.5)));                                  \
-        EXPECT_FALSE(format::isTypeConversionSafe((T1)(3), (T2)(3)));                                  \
+        EXPECT_FALSE(utils::isTypeConversionSafe((T1)(1.5), (T2)(1.5)));                                  \
+        EXPECT_FALSE(utils::isTypeConversionSafe((T1)(3), (T2)(3)));                                  \
       if constexpr (std::is_signed_v<T2>){\
-        EXPECT_FALSE(format::isTypeConversionSafe((T1)(-1.5), (T2)(-1.5)));                                  \
-        EXPECT_FALSE(format::isTypeConversionSafe((T1)(-3), (T2)(-3)));                                  \
+        EXPECT_FALSE(utils::isTypeConversionSafe((T1)(-1.5), (T2)(-1.5)));                                  \
+        EXPECT_FALSE(utils::isTypeConversionSafe((T1)(-3), (T2)(-3)));                                  \
       } else {                        \
-        EXPECT_FALSE(format::isTypeConversionSafe((T1)(-3), (T2)(-3)));                                  \
+        EXPECT_FALSE(utils::isTypeConversionSafe((T1)(-3), (T2)(-3)));                                  \
       }                                  \
     }                                     \
   }                                   \
   if constexpr (std::is_floating_point_v<T2> == false) {                                     \
     if constexpr (std::is_integral_v<T1> == true && std::is_integral_v<T2> == true) {                                    \
       if constexpr (std::is_floating_point_v<T1> == false && std::uintmax_t(std::numeric_limits<T1>::max()) > std::uintmax_t(std::numeric_limits<T2>::max())) {  \
-      EXPECT_FALSE(format::isTypeConversionSafe((T1)(std::numeric_limits<T2>::max() + (unsigned long long)1), (T2)((unsigned long long)std::numeric_limits<T2>::max() + (unsigned long long)1)));} \
+      EXPECT_FALSE(utils::isTypeConversionSafe((T1)(std::numeric_limits<T2>::max() + (unsigned long long)1), (T2)((unsigned long long)std::numeric_limits<T2>::max() + (unsigned long long)1)));} \
       {                                   \
         const intmax_t botT1 = std::numeric_limits<T1>::min(); \
         const intmax_t botT2 = std::numeric_limits<T2>::min(); \
         if (botT1 < botT2) {            \
           if constexpr (!std::is_same_v<signed long long, T2>)                              \
-            EXPECT_FALSE(format::isTypeConversionSafe((T1)(botT2 - 1), (T2)(botT2 - 1)));}\
+            EXPECT_FALSE(utils::isTypeConversionSafe((T1)(botT2 - 1), (T2)(botT2 - 1)));}\
       }\
     }                                   \
   }                                    \
   if constexpr (!std::is_integral_v<T1> && !std::is_integral_v<T2>) {\
     if constexpr (double(std::numeric_limits<T1>::max()) > double(std::numeric_limits<T2>::max())) {  \
-    EXPECT_FALSE(format::isTypeConversionSafe(((T1)(T1)std::numeric_limits<T2>::max() * (T1)2), (T2)(std::numeric_limits<T2>::max() * (T1)2)));} \
+    EXPECT_FALSE(utils::isTypeConversionSafe(((T1)(T1)std::numeric_limits<T2>::max() * (T1)2), (T2)(std::numeric_limits<T2>::max() * (T1)2)));} \
       const double botT1 = double(-(std::numeric_limits<T1>::max())); \
       const double botT2 = double(-(std::numeric_limits<T2>::max())); \
       if constexpr (!std::is_same_v<double, T2>)                              \
         if (botT1 < botT2) {      \
-          EXPECT_FALSE(format::isTypeConversionSafe((T1)(botT2 * 2), (T2)(botT2 * 2)));}                         \
+          EXPECT_FALSE(utils::isTypeConversionSafe((T1)(botT2 * 2), (T2)(botT2 * 2)));}                         \
    }
   //} else if constexpr (std::is_integral_v<T1> && !std::is_integral_v<T2>) {\
   //  if constexpr (double(std::numeric_limits<T1>::max()) > double(std::numeric_limits<T2>::max())) {  \
